Update to use a custom deploy job.
The jenkinslib lambda deploy job is set up to deploy the lambdas for the
backend checks, so it runs the end to end tests among other things and
you can't run the end to end tests against the mgmt environment because
there isn't one. It seems a bit much to update the jenkinslib deploy job
to work with a mgmt account parameter for this one lambda so I've
updated this to use a custom deploy job.