Filter vulnerabilities using scan findings report
When an ECR scan event arrives, fetch the names of the
vulnerabilities from the AWS API. Compare them with a list
of muted vulnerabilities, and only send an alert for ones
which are not muted.

This helps us in the case where a low-severity vulnerability
has been reported, but no fix is available yet. We can mute
the vulnerability until the fix is available so that if any
further alerts are sent, we know they need to be investigated.