Add a description to all tests
This makes the tests names clearer. For example:

```
should not send a Slack message for an ECR scan of 'intg' with no results
```

rather than:

```
should not send a slack message for event
{
  "detail": {
    "scan-status": "COMPLETE",
    "repository-name": "yara-dependencies",
    "image-tags" : ["intg"],
    "finding-severity-counts": {
      "CRITICAL": 0,
      "HIGH": 0,
      "MEDIUM": 0,
      "LOW": 0
    }
  }
}
```

This is particularly useful when a test fails, because IntelliJ can't navigate
to the source of a parameterized test, so you couldn't work out which test case
failed without interpreting the event yourself.