Filter out informational and undefined ECR vulnerabilities
The full scan finding report sometimes includes vulnerabilities at
an 'informational' and 'undefined' levels.

We don't need to act on the informational ones urgently, so don't send alerts.

Also don't send alerts for undefined alerts, so as not to change the behaviour
from before. We may revisit this if we think we should receive alerts for
undefined vulnerabilities.