Edit export slack message to make difference between fail and success clearer
I've added emojis that differ based on fail/success of export.
I've also put in markdown to make the message a little easier to read and get information from at a glance.

From research online, it appears that message colouring is only available when using an app to post messages (such as Jenkins)
and colour cannot be included when sending the message as a section as we are.