Add notification for disk alert messages
There is a cloudwatch alarm set up which monitors the disk usage of the
Jenkins instances.

When the disk usage goes over a threshold or when it goes below the
threshold, a message is sent to slack.

The message contains the server (Jenkins or JenkinsProd), the threshold
and the either a tick or a warning symbol depending on where it's ALARM
or OK.