Create a separate alert message for no data points.
The cloudwatch alarm for Jenkins is set to show an error if it stops
receiving data from the Jenkins Cloudwatch agent. This is good in that
we get alerted if there is a problem with the agent but it's confusing
in that it says that the disk space is above 70% when it isn't.
I've changed this to put in a more sensible message if the problem is
missing data points. The only way is to check for a string in the reason
field but that's unlikely to change so we should be ok.