Add a link to the clear disk space documentation if the disk space is too high
If the alarm is because the disk space on Jenkins is too high, the slack
message has a link to the dev documentation about how to clear it. I've
not put this on the missing data or ok messages as it's not really
needed as there's nothing to do.

I've also put the Grafana url as a proper link with text to make it look
a bit better in the message.