Use transferring body to prevent sqs message send
The end-to-end tests use 0 byte files to test uploading, this causes issues for the Transform Engine. In reality 0 byte files would not be accepted by TDR

To provide control which transfers in the lower environments (integration and staging) use the transferring body name to filter the messages.

The end-to-end tests will use "mock" body names for their transfers. Using the body name will also allow developers and others to control whether a message is sent to the Transform Engine when using the lower environments.