Make a generic message event.
Quite often we just need to send a string of text to slack and there's
no need for the post processing that the notifications service does.

This gets rid of the secret rotation decoder and replaces it with a
generic message sender.

We can probably refactor some of the existing decoders to use this as
well but that can come later.